requirements_diagnostics.txt
# Requirements for VESPA Dashboard Diagnostic Tools
requests>=2.28.0
python-dotenv>=1.0.0
pathlib
